+++ CHANGELOG.md
@@ -1,5 +1,23 @@
 # linkedin-extension
 
+## 0.0.7
+
+### Patch Changes
+
+- a79e81f: fix: response from content.ts
+
+## 0.0.6
+
+### Patch Changes
+
+- 89f7b17: feat: add track profile feature for new connections and DTM
+
+## 0.0.5
+
+### Patch Changes
+
+- ad9c9b3: feat: add shadcn ui components with tailwind v4
+
 ## 0.0.4
 
 ### Patch Changes